Lambda Function (Microservice, Sanic)/requirements.txt
sanic
mysql-connector-python
mangum
uvicorn
pycognito
pyjwt
python-dotenv
sanic_cors